cc65 upgrade fixes; ~ (.BITNOT) results cast with < (.HIBYTE)
In cc65 2.13.3 it was valid to use e.g. #~1 in an 8-bit context, but in
cc65 revision fed3179d3b180ca05ca5ad5037b06e7612d5459f it is rejected
as out of range unless cast with < or .LOBYTE e.g. #<~1

Example code and error:

mask_miso  = %10000000
  AND #~(mask_miso)

src/ili9340.s(144): Error: Range error (18446744073709551487 not in [0..255])